backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
pydantic==2.8.0
pydantic-settings>=2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.1
aiosqlite==0.20.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
bcrypt>=4.0,<5.0
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.9.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
ollama>=0.1.0
opencv-python-headless>=4.9.0
pillow>=10.0.0
aiofiles>=23.0.0
pyyaml>=6.0
